requirements.txt


mcp>=1.0.0
fastmcp>=1.0.0


# --- Google Calendar API ---
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# --- Optional: for MCP + Ollama client integration ---
llama-index-core
llama-index-llms-ollama
llama-index-llms-groq
llama-index-tools-mcp
nest_asyncio
python-dotenv

dateparser
asyncio

# --- Telegram Bot Integration ---
python-telegram-bot

# --- Flask WhatsApp Integration ---
flask
requests
aiohttp
nest-asyncio